Update Design Paper (#61)
Redacted the first half (up to Design requirements). Not perfect, but I think a good step forward to me to focus on the next part / further papers and work with you other writings. I suggest to to an intensive check / edit of this later, when we have finished more things.