Update ERC5XX0 draft.md
Removed these text
(may add it later to a medium article


The `scriptURI` points to decentralized immutable locations *or* the location itself contains information that can validate the script's authenticity. For example, the URI may point to a location on a blockchain, or, URI may contain a hash digest of the resource, such as with IPFS. In the first situation, the client (wallet) should assume it is authentic. In the latter case, the client must also validate that hash digest.

If the `scriptURI` is a download link to a server, it may not contain a digest in the URI. In such a case, it may be desirable to return the digest in a struct together with the URI. This allows the caller to assert the client script is authentic after downloading it. However, igital Signature is a much better way to solve this problem.